Modules refactoring, simplified JavaRedis, documented APIs, restricted visibility
