+++ src/main/scala/play/api/cache/redis/impl/AsyncRedis.scala
@@ -5,8 +5,10 @@ import javax.inject.{Inject, Singleton}
 import play.api.cache.redis._
 
 /**
+  * Implementation of **asynchronous** Redis API
+  *
   * @author Karel Cemus
   */
 @Singleton
-class AsyncRedis @Inject( )( redis: RedisConnector, settings: ConnectionSettings )
+private[ impl ] class AsyncRedis @Inject( )( redis: RedisConnector, settings: ConnectionSettings )
   extends RedisCache( redis, settings )( Builders.AsynchronousBuilder ) with CacheAsyncApi
