+++ src/main/scala/play/api/cache/redis/impl/RedisCache.scala
@@ -8,7 +8,7 @@ import scala.reflect.ClassTag
 import play.api.cache.redis._
 
 /** <p>Implementation of plain API using redis-server cache and Brando connector implementation.</p> */
-class RedisCache[ Result[ _ ] ]( redis: RedisConnector, settings: ConnectionSettings )( implicit builder: Builders.ResultBuilder[ Result ] ) extends InternalCacheApi[ Result ] with Implicits {
+private[ impl ] class RedisCache[ Result[ _ ] ]( redis: RedisConnector, settings: ConnectionSettings )( implicit builder: Builders.ResultBuilder[ Result ] ) extends InternalCacheApi[ Result ] with Implicits {
 
   // implicit execution context and ask timeout
   import settings.{invocationContext, timeout}
