+++ src/main/scala/play/api/cache/redis/impl/SyncRedis.scala
@@ -5,8 +5,10 @@ import javax.inject.{Inject, Singleton}
 import play.api.cache.redis._
 
 /**
+  * Implementation of **synchronous** and **blocking** Redis API. It also implements standard Play Scala CacheApi
+  *
   * @author Karel Cemus
   */
 @Singleton
-class SyncRedis @Inject( )( redis: RedisConnector, settings: ConnectionSettings )
+private[ impl ] class SyncRedis @Inject( )( redis: RedisConnector, settings: ConnectionSettings )
   extends RedisCache( redis, settings )( Builders.SynchronousBuilder ) with CacheApi with play.api.cache.CacheApi
